Ai Agents/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-linkedin-v2==0.8.6
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.3 